git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/ReFlow/branches/gus4@14882 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ StepClasses/lib/perl/MakeDataDir.pm
@@ -23,8 +23,10 @@ sub run {
       
       # otherwise assume regular dir and remove with rmdir
       }else{
-      # should be emtpy
-      $self->runCmd(0, "rmdir $workflowDataDir/$dataDir");
+      # should be empty - check it still exists
+        if (-e "$workflowDataDir/$dataDir"){
+            $self->runCmd(0, "rmdir $workflowDataDir/$dataDir");
+        }
       }
 
   }else{
